fix: drop (long) cast in amx_BrowseRelocate NULL assignment
userdata[] is void*[]; the (long) cast triggers -Wint-conversion under
clang-11 (error with -Werror), while gcc only warns. The cast is a
2005-era stylistic mistake — both forms produce a null void* on every
target we build (i386 LP32, amd64/aarch64 LP64, Windows x64 LLP64).

Note the assignment is gated on the 64-bit-cell branch (PAWN_CELL_SIZE
widening on 64bit), so upstream master never compiles this line — it
sits inside the original PAWN_CELL_SIZE==16 block. The widened gate on
64bit brings it into amd64/aarch64 compilation, exposing the cast.

+++ amxmodx/amx.cpp
@@ -571,7 +571,7 @@ static int amx_BrowseRelocate(AMX *amx)
      */
     if (sizeof(AMX_NATIVE)<=sizeof(cell))
       amx->sysreq_d=OP_SYSREQ_D;
-	amx->userdata[UD_OPCODELIST] = (long)NULL;
+	amx->userdata[UD_OPCODELIST] = NULL;
   #endif
 
   /* start browsing code */
